refactor: uploadImage 로직 utils로 분리
